podcast_service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.3.0
httpx>=0.27.2
google-cloud-texttospeech>=2.14.1
google-cloud-storage>=2.18.0
pydub>=0.25.1
pyyaml>=6.0.1
requests>=2.31.0
